Screnshots of the entire website
This png contains all the pages of the website with the work done. For example, in the table area, is a little bit of basic PHP with some numerical operations. On the Contact page is a Contact form made with the "Contact-7 form". Some pages are created for test reasons. Dear reader, hope you enjoy my work. It is my first website, created from scratch.